Refactor ActivityHelper
