+++ app/src/test/java/org/mercycorps/translationcards/util/AddTranslationActivityHelper.java
@@ -1,0 +1,7 @@
+package org.mercycorps.translationcards.util;
+
+/**
+ * Created by agarrard on 5/17/16.
+ */
+public class AddTranslationActivityHelper {
+}
